# frontend/landings/agivys/landing/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgiVysSystem API",
    "description": "API para gestão do ecossistema AgiVysSystem",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://joederblanca.com.br/agivys-api"
    }
  ],
  "paths": {
...
    "/api/v1/address/my-address": {
      "get": {
        "tags": [
          "Address"
        ],
        "summary": "Listagem de Endereços do Usuário",
        "description": "Recupera todos os endereços físicos vinculados ao perfil pessoal (Person) do usuário atualmente autenticado.\r\nÉ necessário enviar o Bearer Token no cabeçalho da requisição.",
        "responses": {
          "200": {
            "description": "Lista de endereços recuperada com sucesso."
          },
          "401": {
            "description": "Usuário não autenticado ou token inválido."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "post": {
        "tags": [
          "Address"
        ],
        "summary": "Cadastro de Novo Endereço",
        "description": "Adiciona um novo endereço físico ao perfil do usuário autenticado. \r\nO usuário deve possuir um registro de `Person` previamente criado no banco de dados.",
        "requestBody": {
          "description": "Dados completos do endereço, incluindo CEP, Logradouro, Número, Bairro, Cidade e Estado.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPersonDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPersonDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPersonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Endereço cadastrado com sucesso."
          },
          "400": {
            "description": "Dados inválidos ou erro de validação."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "404": {
            "description": "Perfil (Person) não localizado para o usuário autenticado."
          },
          "500": {
            "description": "Erro interno ao processar a inclusão."
          }
        }
      }
    },
    "/api/v1/address/my-address/{id}": {
      "put": {
        "tags": [
          "Address"
        ],
        "summary": "Atualização de Endereço Existente",
        "description": "Modifica os dados de um endereço previamente cadastrado.\r\n**Regra de Segurança**: O endereço a ser alterado deve pertencer estritamente ao perfil do usuário autenticado. Tentativas de alterar endereços de terceiros resultarão em bloqueio de acesso (HTTP 403).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID único do endereço a ser atualizado.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Objeto contendo os novos dados do endereço.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPersonDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPersonDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPersonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Endereço atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos no corpo da requisição."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Permissão negada. O endereço pertence a outro usuário."
          },
          "404": {
            "description": "Endereço não localizado na base de dados."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "delete": {
        "tags": [
          "Address"
        ],
        "summary": "Exclusão de Endereço",
        "description": "Remove de forma permanente um endereço vinculado ao perfil do usuário autenticado.\r\n**Regra de Segurança**: A exclusão só será permitida caso o endereço pertença ao usuário logado, protegendo os dados contra acesso indevido (Insecure Direct Object Reference).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID único numérico do endereço a ser removido.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Endereço removido permanentemente com sucesso."
          },
          "401": {
            "description": "Usuário não autenticado."
          },
          "403": {
            "description": "Ação negada. O endereço não pertence ao usuário."
          },
          "404": {
            "description": "O endereço especificado não foi encontrado."
          },
          "500": {
            "description": "Erro interno do servidor ao processar a exclusão."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressPersonDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "complement": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Cole APENAS o token JWT.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}